# perf: optimize tree resolve