--- a/pkg/model/stacktraces.go
+++ b/pkg/model/stacktraces.go
@@ -213,6 +213,7 @@ func (t *StacktraceTree) LookupLocations(dst []uint64, idx int32) []uint64 {
 
 // MinValue returns the minimum "total" value a node in a tree has to have.
 func (t *StacktraceTree) MinValue(maxNodes int64) int64 {
+	// TODO(kolesnikovae): Consider quickselect.
 	if maxNodes < 1 || maxNodes >= int64(len(t.Nodes)) {
 		return 0
 	}
@@ -286,8 +287,6 @@ func (t *StacktraceTree) Traverse(maxNodes int64, fn StacktraceTreeTraverseFn) e
 	return nil
 }
 
-var lostDuringSerializationNameBytes = []byte(truncatedNodeName)
-
 func (t *StacktraceTree) Bytes(dst io.Writer, maxNodes int64, funcs []string) {
 	if len(t.Nodes) == 0 || len(funcs) == 0 {
 		return
@@ -302,9 +301,8 @@ func (t *StacktraceTree) Bytes(dst io.Writer, maxNodes int64, funcs []string) {
 			// and the byte slice backing capacity is managed by GC.
 			name = unsafeStringBytes(funcs[n.Location])
 		case sentinel:
-			name = lostDuringSerializationNameBytes
+			name = truncatedNodeNameBytes
 		}
-
 		_, _ = vw.Write(dst, uint64(len(name)))
 		_, _ = dst.Write(name)
 		_, _ = vw.Write(dst, uint64(n.Value))
@@ -315,4 +313,42 @@ func (t *StacktraceTree) Bytes(dst io.Writer, maxNodes int64, funcs []string) {
 
 func unsafeStringBytes(s string) []byte {
 	return unsafe.Slice(unsafe.StringData(s), len(s))
+}
+
+func (t *StacktraceTree) Tree(maxNodes int64, names []string) *Tree {
+	if len(t.Nodes) < 2 || len(names) == 0 {
+		// stack trace tree has root at 0: trees with less
+		// than 2 nodes are considered empty.
+		return new(Tree)
+	}
+
+	nodesSize := maxNodes
+	if nodesSize < 1 || nodesSize > 10<<10 {
+		nodesSize = 1 << 10 // Sane default.
+	}
+	root := new(node) // Virtual root node.
+	nodes := make([]*node, 1, nodesSize)
+	nodes[0] = root
+	var current *node
+
+	_ = t.Traverse(maxNodes, func(index int32, children []int32) error {
+		current, nodes = nodes[len(nodes)-1], nodes[:len(nodes)-1]
+		sn := &t.Nodes[index]
+		var name string
+		if sn.Location < 0 {
+			name = truncatedNodeName
+		} else {
+			name = names[sn.Location]
+		}
+		n := current.insert(name)
+		n.self = sn.Value
+		n.total = sn.Total
+		n.children = make([]*node, 0, len(children))
+		for i := 0; i < len(children); i++ {
+			nodes = append(nodes, n)
+		}
+		return nil
+	})
+
+	return &Tree{root: root.children[0].children}
 }