--- a/pkg/model/tree.go
+++ b/pkg/model/tree.go
@@ -129,6 +129,14 @@ func (t *Tree) IterateStacks(cb func(name string, self int64, stack []string)) {
 const defaultDFSSize = 128
 
 func (t *Tree) Merge(src *Tree) {
+	if t.Total() == 0 && src.Total() > 0 {
+		*t = *src
+		return
+	}
+	if src.Total() == 0 {
+		return
+	}
+
 	srcNodes := make([]*node, 0, defaultDFSSize)
 	srcRoot := &node{children: src.root}
 	srcNodes = append(srcNodes, srcRoot)
@@ -325,6 +333,8 @@ func (h *minHeap) Pop() interface{} {
 }
 
 const truncatedNodeName = "other"
+
+var truncatedNodeNameBytes = []byte(truncatedNodeName)
 
 // MarshalTruncate writes tree byte representation to the writer provider,
 // the number of nodes is limited to maxNodes. The function modifies