--- a/pkg/phlaredb/symdb/resolver_tree.go
+++ b/pkg/phlaredb/symdb/resolver_tree.go
@@ -1,6 +1,7 @@
 package symdb
 
 import (
+	"bytes"
 	"sync"
 
 	"github.com/grafana/pyroscope/pkg/model"
@@ -8,11 +9,12 @@ import (
 )
 
 type treeSymbols struct {
-	symbols *Symbols
-	samples *schemav1.Samples
-	tree    *model.Tree
-	lines   []string
-	cur     int
+	symbols  *Symbols
+	samples  *schemav1.Samples
+	tree     *model.StacktraceTree
+	maxNodes int64
+	lines    []int32
+	cur      int
 }
 
 var treeSymbolsPool = sync.Pool{
@@ -26,7 +28,7 @@ func treeSymbolsFromPool() *treeSymbols {
 func (r *treeSymbols) reset() {
 	r.symbols = nil
 	r.samples = nil
-	r.tree = nil
+	r.tree.Reset()
 	r.lines = r.lines[:0]
 	r.cur = 0
 	treeSymbolsPool.Put(r)
@@ -35,18 +37,32 @@ func (r *treeSymbols) reset() {
 func (r *treeSymbols) init(symbols *Symbols, samples schemav1.Samples) {
 	r.symbols = symbols
 	r.samples = &samples
-	r.tree = new(model.Tree)
+	if r.tree == nil {
+		// Branching factor.
+		r.tree = model.NewStacktraceTree(samples.Len() * 2)
+	}
 }
 
 func (r *treeSymbols) InsertStacktrace(_ uint32, locations []int32) {
 	r.lines = r.lines[:0]
-	for i := len(locations) - 1; i >= 0; i-- {
+	for i := 0; i < len(locations); i++ {
 		lines := r.symbols.Locations[locations[i]].Line
-		for j := len(lines) - 1; j >= 0; j-- {
+		for j := 0; j < len(lines); j++ {
 			f := r.symbols.Functions[lines[j].FunctionId]
-			r.lines = append(r.lines, r.symbols.Strings[f.Name])
+			r.lines = append(r.lines, int32(f.Name))
 		}
 	}
-	r.tree.InsertStack(int64(r.samples.Values[r.cur]), r.lines...)
+	r.tree.Insert(r.lines, int64(r.samples.Values[r.cur]))
 	r.cur++
+}
+
+func (r *treeSymbols) buildTree() *model.Tree {
+	// TODO(kolesnikovae): Eliminate intermediate serialization.
+	var buf bytes.Buffer
+	r.tree.Bytes(&buf, r.maxNodes, r.symbols.Strings)
+	t, err := model.UnmarshalTree(buf.Bytes())
+	if err != nil {
+		panic(err)
+	}
+	return t
 }